Move MatchingFileIterator to its own file for phpcs
